+++ internal/controller/vspherevalidator_controller.go
@@ -37,11 +37,11 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/reconcile"
 
 	"github.com/validator-labs/validator-plugin-vsphere/api/v1alpha1"
-	"github.com/validator-labs/validator-plugin-vsphere/internal/constants"
-	"github.com/validator-labs/validator-plugin-vsphere/internal/validators/computeresources"
-	"github.com/validator-labs/validator-plugin-vsphere/internal/validators/ntp"
-	"github.com/validator-labs/validator-plugin-vsphere/internal/validators/privileges"
-	"github.com/validator-labs/validator-plugin-vsphere/internal/validators/tags"
+	"github.com/validator-labs/validator-plugin-vsphere/pkg/constants"
+	"github.com/validator-labs/validator-plugin-vsphere/pkg/validators/computeresources"
+	"github.com/validator-labs/validator-plugin-vsphere/pkg/validators/ntp"
+	"github.com/validator-labs/validator-plugin-vsphere/pkg/validators/privileges"
+	"github.com/validator-labs/validator-plugin-vsphere/pkg/validators/tags"
 	"github.com/validator-labs/validator-plugin-vsphere/pkg/vsphere"
 	vapi "github.com/validator-labs/validator/api/v1alpha1"
 	"github.com/validator-labs/validator/pkg/types"
